✨ Dynamic topology validation
Detects "kinks" in polygons while user digitizes. I'm struggling to get this working reliably. mapbox-gl-draw for some reason has it's own render loop and it causes issues with setting properties on shapes to drive changes in symbology as topology issues are created. This should be workable but this branch will need some more debugging to work well enough to deploy.