Offline App Shell routing and SW prefetching
App Shell refers to the small index.html page that is served for *all* navigable routes in the application. A ServiceWorker route handler ensures that the single cached index.html can be used when offline for all routes. First the network is checked for a more up-to-date response.

This commit also includes an install event handler in the SW. This will trigger prefetch caching of all application code if that feature has been enabled by the user.