Improve SWR implementation in GraphqlQueryCache
Previous implementation of stale-while-revalidate had the service-worker notify the main window to refetch queries. This posed the risk of causing an infinite loop that was mitigated using "pauseRevalidation" flags in the service worker. The whole thing seemed risky and brittle so instead this implementation just writes to the apollo cache directly when data is revalidated. No risk of looping through swr over and over.