💄 Replacing alert, prompt, and confirm native calls
Now using custom components based on ModalRevisited, which will soon replace all ModalDeprecated components.

In many cases I had to weigh creating a custom dialog to achieve UI goals vs just using the native components. With this change it will be much easier to just bring in `useDialog` to get a more customizable look. The native components are also being [deprecated in web standards](https://github.com/whatwg/html/issues/2894) so this is a move in the right direction.