Basic sketch geometry editing persistence
TODO: caching is not implemented at all, so each time a sketch is shown it has to be refetched.

TODO: There can be a (potentially long) period where the sketch isn't shown between submitting edits and going back to the ToC. This is because the MapContextManager refetches the sketch GeoJSON. Will need to do something more sophisticated where it keeps recently edited sketches in a cache of some sort.