Improvements to Overlay Admin toolbar
Now a proper "file menu" style toolbar in anticipation of additional feature additions that would over-complicated the previous design.

The _Publish_ button now shows an semi-disabled appearance when there are no changes to publish.

There is now a dropdown menu action for uploading spatial data that shows a file picker. It limits selections to .tif, .geojson, .json, .fgb, and .zip