Performance improvements to duplicate_ip checking
When submitting a survey response, we store a hash of the request ip if available to detect repeated entries and potential abuse. IP addresses are hashed to protect privacy but checking hashes is slow, and becomes slower as the number of records grows. This is a fundamental limitation since we must store IP addresses with unique salts to protect privacy.

To improve performance, we will now keep a maximum of 1000 hash records and also otherwise age them out after 1 week.