Data library (#834)
Adds support for the "Data Library", which will be a growing curated collection of authoritative data layers which are easy to add to any project. First cut at this will be focused on publishing a tree of table of contents items (layers) from the superuser project, which get updated on a periodic basis. First case study is NOAA Coral Reef Watch layers.